refactor: Rename mopidy.{_server => _app}
