+++ src/mopidy/__main__.py
@@ -1,6 +1,6 @@
 import sys
 
-from mopidy._server.main import main
+from mopidy._app.main import main
 
 if __name__ == "__main__":
     sys.exit(main())
